// src/app/shared/generated/model/imageDataResponse.ts
/**
 * onecx-welcome-bff
 * welcome bff
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: tkit_dev@1000kit.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ImageDataResponse { 
    imageId?: string;
    imageData?: Blob;
    dataLength?: number;
    mimeType?: string;
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
}